requirements.txt
# Core dependencies
pandas==1.3.5  # For data manipulation and Excel file reading
PyYAML==6.0    # For YAML generation and manipulation
